Rename pkg stats -> statistics
