+++ internal/pkg/service/buffer/statistics/manager.go
@@ -1,4 +1,4 @@
-package stats
+package statistics
 
 import (
 	"context"
