+++ internal/pkg/service/buffer/statistics/manager_test.go
@@ -1,4 +1,4 @@
-package stats
+package statistics
 
 import (
 	"context"
